+++ .gitignore
@@ -44,4 +44,6 @@ local.properties
 buck-out/
 \.buckd/
 *.keystore
-      
+
+index.js
+index.d.ts
